Organised tests into separate files
